chore(origin-backend): Use numbers instead of Role in DTOs, move org DTOs to a separate folder
